Move dropdown to complex elements
