+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/elements/complex/Dropdown.java
@@ -1,4 +1,4 @@
-package com.epam.jdi.light.ui.bootstrap.elements.common;
+package com.epam.jdi.light.ui.bootstrap.elements.complex;
 
 import com.epam.jdi.light.asserts.generic.UIAssert;
 import com.epam.jdi.light.elements.base.UIBaseElement;
@@ -6,6 +6,7 @@ import com.epam.jdi.light.elements.common.UIElement;
 import com.epam.jdi.light.elements.complex.WebList;
 import com.epam.jdi.light.elements.interfaces.common.IsButton;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.Css;
+import com.epam.jdi.light.ui.bootstrap.elements.common.Button;
 
 public class Dropdown extends UIBaseElement<UIAssert> implements IsButton{
     @Css(".dropdown-toggle") public Button dropdownToggle;
